+++ ReleaseNotes.md
@@ -1,5 +1,16 @@
 # Release Notes
 
+## Version 0.0.11
+
+* OpenAPI: 0.0.11
+* Platform API Server: 0.0.11
+
+### Features
+* **GET /about.json**
+  - retrieve about information
+* **POST/PATCH Organization**
+  - optional: accepts different tokens for solace cloud & event portal
+
 ## Version 0.0.10
 
 Initial release.